fix: fixing linting
